requirements.txt
environs==10.3.0
python-telegram-bot==13.15
redis==5.0.1
vk_api==11.9.9
argparse==1.4.0
